Create HowToIgnoreFile_JAVA_JBOSS_Android
This file have to be place at the root of the project. Also it has to renamed for GIT to .gitignore, or for SVN to svn:ignore.
In TurtoiseSVN can be set as global ignore list by right click on Desktop, selecting TurtoiseSVN--Setings, where you can add this in addition to existing ignores as ONE Line String, such as: "*.class build/* j2meclasses/* .classpath .project .settings *.war *~ *.lock *.DS_Store *.swp *.out *.bak *.tmp *.log log*.txt .idea .git" (exclude quotes)
The SECOND METHOD is: in the Repo-Browser right click on folder and select "Show properties" when select "New" then"Other",type in "svn:ignore" for the group name and copy/paste or manually entry your ignores and click "OK"
This is it!